refactor: annotation to typescript
